Factor out sample groupings
